Separate the 'waitForIndexer' function into separate module usable by other tests
